prod

+++ src/start.py
@@ -17,17 +17,14 @@ initial_config = Dynaconf(
 
 def main():
     logger.info(f'Numero de pool {initial_config.N_POOL}')
-    #if downloads_files():
-    clear_tables()
-    tifs_path = f'{settings.CATALOG}cempa_tifs'
-    if isdir(tifs_path):
-        rmtree(tifs_path)
-    mkdir(tifs_path)
-        
-    to_db()
-    #else:
-    #    pass
-    #main(initial_config.FORCE_SAVE_BD)
+    if downloads_files():
+        clear_tables()
+        tifs_path = f'{settings.CATALOG}cempa_tifs'
+        if isdir(tifs_path):
+            rmtree(tifs_path)
+        mkdir(tifs_path)
+        to_db()
+    else:
 
 
 if __name__ == '__main__':
